Refactor cli script to typescript
