+++ package.json
@@ -28,6 +28,7 @@
     "@types/geojson": "^7946.0.16"
   },
   "devDependencies": {
+    "@types/node": "^24.12.2",
     "@vitest/coverage-v8": "^4.1.5",
     "rolldown": "^1.0.0-rc.18",
     "rolldown-plugin-dts": "^0.24.1",
@@ -35,7 +36,7 @@
     "vitest": "^4.1.5"
   },
   "bin": {
-    "osm2geojson-lite": "cli.mjs"
+    "osm2geojson-lite": "dist/cli.js"
   },
   "directories": {
     "lib": "lib",
